refactor: split into lib + bin with separate test file
Move plugin logic from src/main.rs to src/lib.rs and tests to
src/tests.rs, following the same lib+bin pattern as zjstatus.

- src/lib.rs: all plugin logic (State, ZellijPlugin impl)
- src/main.rs: just register_plugin!() entry point
- src/tests.rs: all unit tests with FFI stub
- CI workflows updated to use --lib for tests and clippy
