+++ Cargo.toml
@@ -4,6 +4,14 @@ version = "0.1.0"
 edition = "2021"
 license = "MIT"
 
+[lib]
+name = "zellij_attention"
+path = "src/lib.rs"
+
+[[bin]]
+name = "zellij-attention"
+path = "src/main.rs"
+
 [dependencies]
 zellij-tile = "0.43.1"
 serde = { version = "1", features = ["derive"] }
